pyproject.toml
[tool.poetry]
name = "flax-jl"
version = "0.1.0"
description = ""
authors = ["vincentmolin <molinv@chalmers.se>"]
license = "MIT"
readme = "README.md"
packages = [{include = "flax"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
jax = "^0.3"
numpy = "^1"
flax = "^0.6.0"
scipy = "^1.9.1"
dm-haiku = "^0.0.7"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"
ipykernel = "^6.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
